Fix lucene-query-access export in src/index.ts
