+++ packages/xlucene-evaluator/src/index.ts
@@ -1,13 +1,11 @@
 import DocumentMatcher from './document-matcher';
 import LuceneQueryParser from './lucene-query-parser';
-import LuceneQueryAccess, { QueryAccessConfig } from './lucene-query-access';
 
+export * from './lucene-query-access';
 export * from './interfaces';
 export * from './utils';
 export * from './translator';
 export {
     DocumentMatcher,
     LuceneQueryParser,
-    LuceneQueryAccess,
-    QueryAccessConfig,
 };
